+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     - Type
     - Cases
 - Add `KeyWord` `Do` and parse it as a `Parsed.KeyWordOnly`
+- Add a field with `DomainType` to `ParticipantComponent`
 
 ## 2.0.1 - 2020-11-06
 - Add missing `KeyWord` value to `KeyWordIf`